extracted hook to utils file
