+++ src/components/Button/Button.stories.mdx
@@ -1,6 +1,7 @@
 import { Meta, Story, ArgsTable, Canvas } from '@storybook/addon-docs/blocks'
 import { boolean, select, text } from '@storybook/addon-knobs'
-import Button, { SIZES, THEMES } from './Button'
+import Button from './Button'
+import { SIZES, THEMES } from './Button.utils'
 import { ControlGroup } from '../index'
 
 <Meta
